backend/requirements.txt
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.7.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
CurrencyConverter==0.17.15
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.3
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-filter==25.1
django-tinymce==4.1.0
djangorestframework==3.16.0
executing==2.0.1
fastjsonschema==2.21.1
Flask==3.0.0
fonttools==4.46.0
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.28.0
ipython==8.19.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.6.0
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.2
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
prometheus_client==0.22.1
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==3.3.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
seaborn==0.13.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.12
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
whitenoise==6.9.0
